src/config/settings/__init__.py
# -*- coding: utf-8 -*-
from .local import Local
from .runtime import Runtime
from .test import Test
from .unit_test import UnitTest
from dotenv import load_dotenv, find_dotenv


load_dotenv(find_dotenv())
