src/requirements/runtime.txt
cython==0.29.28
gevent==21.12.0
pytz==2022.1
cffi==1.15.0
psycopg2-binary==2.9.3
redis==4.2.2
boto3==1.20.37
olefile==0.46
pillow==9.1.0
phonenumbers==8.12.40
attrs==21.4.0
requests[security,socks]==2.27.1
pyamqp==0.1.0.7
celery[tblib]==5.2.6
bcrypt==3.2.0
hiredis==2.0.0
python-dateutil==2.8.2
django==3.2.13
gunicorn==20.1.0
django-configurations==2.3.2
djangorestframework==3.13.1
djangorestframework-filters==1.0.0.dev2
django-redis==5.2.0
django-celery-email==3.0.0
django-templated-email==3.0.0
django-storages==1.12.3
django-cors-headers==3.11.0
django-watchman==1.3.0
django-ratelimit==3.0.1
django-rosetta==0.9.8
django-celery-beat==2.2.1
sorl-thumbnail==12.8.0
collectfast==2.2.0
single-beat==0.5.1
sentry-sdk==1.5.6
django-countries==7.3.2
django-constance==2.8.0
django-picklefield==3.0.1
stripe==2.74.0
dj-stripe==2.6.1
SQLAlchemy==1.4.36
PyMySQL==1.0.2
djangorestframework-camel-case==1.3.0
django-cms==3.10.0
django-sekizai==3.0.1
django-filer==2.2
djangocms-text-ckeditor==5.0.1
djangocms-link==3.0.0
djangocms-picture==4.0.0
reportlab==3.6.11
svglib==1.3.0
whitenoise
django-newsfeed
django-health-check
python-dotenv
